# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Family Chores API
  version: 1.0.0
  description: API for managing family chores, todos, shopping lists, and settings
servers:
  - url: http://localhost:3000/api
    description: Development server
paths:
  /auth/register:
    post:
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                  error:
                    type: string
                    nullable: true
        '400':
          description: Bad request or email already registered
  /auth/login:
    post:
      summary: Login user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                  error:
                    type: string
                    nullable: true
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                  error:
                    type: string
                    nullable: true
        '401':
          description: Invalid or missing refresh token
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
...